# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "CloudSek Posts & Comments API",
		"version": "1.0.0"
	},
	"servers": [
		{ "url": "http://localhost:3000" }
	],
	"components": {
		"securitySchemes": {
			"XAuthToken": {
				"type": "apiKey",
				"in": "header",
				"name": "x-auth-token"
			}
		},
		"schemas": {
			"Register": {
				"type": "object",
				"required": ["username", "email", "password"],
				"properties": {
					"username": {"type": "string"},
					"email": {"type": "string", "format": "email"},
					"password": {"type": "string", "format": "password"}
				}
			},
			"Login": {
				"type": "object",
				"required": ["email", "password"],
				"properties": {
					"email": {"type": "string", "format": "email"},
					"password": {"type": "string", "format": "password"}
				}
			},
			"PostCreate": {
				"type": "object",
				"required": ["title", "content"],
				"properties": {
					"title": {"type": "string"},
					"content": {"type": "string"}
				}
			},
			"CommentCreate": {
				"type": "object",
				"required": ["text"],
				"properties": {
					"text": {"type": "string"},
					"richText": {"oneOf": [{"type": "string"}, {"type": "object"}], "nullable": true}
				}
			}
		}
	},
	"security": [],
	"paths": {
		"/api/auth/register": {
			"post": {
				"summary": "Register a user",
				"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Register"}}}},
				"responses": {"201": {"description": "Registered"}, "400": {"description": "Bad Request"}}
			}
		},
		"/api/auth/login": {
			"post": {
				"summary": "Login",
				"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Login"}}}},
				"responses": {"200": {"description": "OK"}, "400": {"description": "Invalid"}}
			}
		},
		"/api/protected": {
			"get": {
				"summary": "Protected ping",
				"security": [{"XAuthToken": []}],
				"responses": {"200": {"description": "OK"}, "401": {"description": "Unauthorized"}}
			}
		},
		"/api/posts": {
			"get": {
				"summary": "List posts",
				"parameters": [
					{"name": "skip", "in": "query", "schema": {"type": "integer", "minimum": 0}},
					{"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100}}
				],
				"responses": {"200": {"description": "OK"}}
			},
			"post": {
				"summary": "Create post",
				"security": [{"XAuthToken": []}],
				"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostCreate"}}}},
				"responses": {"201": {"description": "Created"}, "400": {"description": "Bad Request"}, "401": {"description": "Unauthorized"}}
			}
		},
		"/api/posts/{id}": {
			"get": {
				"summary": "Get post by id",
				"parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
				"responses": {"200": {"description": "OK"}, "404": {"description": "Not Found"}}
			},
			"put": {
				"summary": "Update post",
				"security": [{"XAuthToken": []}],
				"parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
				"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostCreate"}}}},
				"responses": {"200": {"description": "OK"}, "401": {"description": "Unauthorized"}, "403": {"description": "Forbidden"}, "404": {"description": "Not Found"}}
			},
			"delete": {
				"summary": "Delete post",
				"security": [{"XAuthToken": []}],
				"parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
				"responses": {"204": {"description": "No Content"}, "401": {"description": "Unauthorized"}, "403": {"description": "Forbidden"}, "404": {"description": "Not Found"}}
			}
		},
		"/api/{postId}/comments": {
			"get": {
				"summary": "List top-level comments for a post",
				"parameters": [
					{"name": "postId", "in": "path", "required": true, "schema": {"type": "string"}},
					{"name": "skip", "in": "query", "schema": {"type": "integer", "minimum": 0}},
					{"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100}}
				],
				"responses": {"200": {"description": "OK"}, "404": {"description": "Post Not Found"}}
			},
			"post": {
				"summary": "Create comment on a post",
				"security": [{"XAuthToken": []}],
				"parameters": [{"name": "postId", "in": "path", "required": true, "schema": {"type": "string"}}],
				"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommentCreate"}}}},
				"responses": {"201": {"description": "Created"}, "400": {"description": "Bad Request"}, "401": {"description": "Unauthorized"}, "404": {"description": "Post Not Found"}}
			}
		},
		"/api/comments/{id}": {
			"put": {
				"summary": "Update a comment",
				"security": [{"XAuthToken": []}],
				"parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
				"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommentCreate"}}}},
				"responses": {"200": {"description": "OK"}, "401": {"description": "Unauthorized"}, "403": {"description": "Forbidden"}, "404": {"description": "Not Found"}}
			},
			"delete": {
				"summary": "Delete a comment",
				"security": [{"XAuthToken": []}],
				"parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
				"responses": {"204": {"description": "No Content"}, "401": {"description": "Unauthorized"}, "403": {"description": "Forbidden"}, "404": {"description": "Not Found"}}
			}
		},
		"/api/comments/{id}/replies": {
			"get": {
				"summary": "List replies for a comment",
				"parameters": [
					{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
					{"name": "skip", "in": "query", "schema": {"type": "integer", "minimum": 0}},
					{"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100}}
				],
				"responses": {"200": {"description": "OK"}, "404": {"description": "Not Found"}}
			},
			"post": {
				"summary": "Create reply to a comment",
				"security": [{"XAuthToken": []}],
				"parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
				"requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommentCreate"}}}},
				"responses": {"201": {"description": "Created"}, "400": {"description": "Bad Request"}, "401": {"description": "Unauthorized"}, "404": {"description": "Not Found"}}
			}
		}
	}
} 